requirements.txt
feedparser~=5.2.1
pandas~=1.2.4
requests~=2.24.0
html2text~=2020.1.16
beautifulsoup4~=4.9.1
pytz~=2020.1
python-dateutil~=2.8.1